+++ modules/light-clients/08-wasm/client/cli/tx.go
@@ -1,6 +1,7 @@
 package cli
 
 import (
+	"encoding/hex"
 	"fmt"
 	"os"
 
@@ -97,14 +98,19 @@ func newMigrateContractCmd() *cobra.Command {
 			}
 
 			clientID := args[0]
-			checksum := args[1]
+			checksumHex := args[1]
 			migrateMsg := args[2]
+
+			checksum, err := hex.DecodeString(checksumHex)
+			if err != nil {
+				return fmt.Errorf("invalid checksum format: %w", err)
+			}
 
 			// Construct the message
 			msg := &types.MsgMigrateContract{
 				Signer:   clientCtx.GetFromAddress().String(),
 				ClientId: clientID,
-				Checksum: []byte(checksum),
+				Checksum: checksum,
 				Msg:      []byte(migrateMsg),
 			}